[refactor]フォルダ階層を整理
