+++ app/src/main/java/com/example/pokebook/ui/screen/Navigation/ScreenRoutes.kt
@@ -1,4 +1,4 @@
-package com.example.pokebook.ui.screen
+package com.example.pokebook.ui.screen.Navigation
 
 /**
  * HOMEタブ
